requirements.txt
# requirements.txt for RF-DETR Project
# Core dependencies
numpy
pandas
scikit-learn
Pillow
matplotlib
pycocotools
# Deep learning
torch
torchvision
# Data augmentation
albumentations
# Jupyter (optional)
jupyter
# For downloading datasets (optional)
kaggle
roboflow
